refactor: move cost_benefit methods for other file and common methods to a video superclass
